requirements.txt
Flask==2.0.1
Pillow>=8.3.2
torch==1.8.1
PyYAML==5.4.1
torchvision==0.9.1
vietocr==0.3.2
python-dotenv==0.20.0